0.1.7-test8

+++ src/main/java/net/lawaxi/command/ShitBoyCommand.java
@@ -1,7 +1,7 @@
 package net.lawaxi.command;
 
 import net.lawaxi.Shitboy;
-import net.lawaxi.util.handler.Pocket48Handler;
+import net.lawaxi.handler.Pocket48Handler;
 import net.mamoe.mirai.console.command.CommandSender;
 import net.mamoe.mirai.console.command.java.JCompositeCommand;
 
@@ -24,7 +24,7 @@ public class ShitBoyCommand extends JCompositeCommand {
     @SubCommand({"getRoomID"})
     public void getRoomID(CommandSender sender, int starID) {
         Pocket48Handler handler = Shitboy.INSTANCE.getHandlerPocket48();
-        for (Integer id : handler.getChannelIDBySeverID(handler.getServerIDByStarID(starID)))
+        for (Long id : handler.getChannelIDBySeverID(handler.getServerIDByStarID(starID)))
             sender.sendMessage(id == null ? "无服务器" : String.valueOf(id));
     }
 }
