+++ src/main/java/net/lawaxi/model/Pocket48Subscribe.java
@@ -3,21 +3,21 @@ package net.lawaxi.model;
 import java.util.List;
 
 public class Pocket48Subscribe {
-    private final List<Integer> roomIDs;
-    private final List<Integer> starIDs;
+    private final List<Long> roomIDs;
+    private final List<Long> starIDs;
     private final boolean showAtOne;
 
-    public Pocket48Subscribe(boolean showAtOne, List<Integer> roomIDs, List<Integer> starIDs) {
+    public Pocket48Subscribe(boolean showAtOne, List<Long> roomIDs, List<Long> starIDs) {
         this.roomIDs = roomIDs;
         this.starIDs = starIDs;
         this.showAtOne = showAtOne;
     }
 
-    public List<Integer> getRoomIDs() {
+    public List<Long> getRoomIDs() {
         return roomIDs;
     }
 
-    public List<Integer> getStarIDs() {
+    public List<Long> getStarIDs() {
         return starIDs;
     }
